+++ src/core/course_hanlder.py
@@ -46,8 +46,10 @@ class CourseHanlder:
     @classmethod
     def del_info(cls, name):
         if not name:
-            raise ValueError('Name not exist')
+            raise ValueError('data not found')
         info = db.session.query(Course).filter(Course.name == name).first()
+        if not info:
+            raise ValueError('Name not exist')
         db.session.delete(info)
         db.session.commit()
         return {'success': True}
@@ -55,8 +57,10 @@ class CourseHanlder:
     @classmethod
     def update_info(cls, name, classroom_id, is_active, open_time):
         if not name:
-            raise ValueError('Name not exist')
+            raise ValueError('data not found')
         info = db.session.query(Course).filter(Course.name == name).first()
+        if not info:
+            raise ValueError('Name not exist')
         Course.classroom_id = classroom_id
         Course.is_active = is_active
         Course.open_time = open_time
