+++ src/core/teacher_hanlder.py
@@ -44,8 +44,10 @@ class TeacherHanlder:
     @classmethod
     def delete_info(cls, name):
         if not name:
-            raise ValueError('User not exist')
+            raise ValueError('data not found')
         user = db.session.query(Teacher).filter(Teacher.name == name).first()
+        if not user:
+            raise ValueError('User not exist')
         db.session.delete(user)
         db.session.commit()
         return {'success': True}
@@ -53,8 +55,10 @@ class TeacherHanlder:
     @classmethod
     def update_info(cls, name, phone_number, gender):
         if not name:
-            raise ValueError('Teacher name not exist')
+            raise ValueError('data not found')
         user = db.session.query(Teacher).filter(Teacher.name == name).first()
+        if not user:
+            raise ValueError('Teacher name not exist')
         user.phone_number = phone_number
         user.gender = gender
         db.session.add(user)
@@ -66,9 +70,4 @@ class TeacherHanlder:
         db.session.commit()
 
         return results
-
-
-
-
-
